+++ test/index.test.js
@@ -9,7 +9,7 @@ const SwaggerParser = require('swagger-parser');
 const swagger = require('../lib');
 
 describe('feathers-swagger', () => {
-  describe('basic functionality', () => {
+  describe('basic functionality with openapi v2', () => {
     let server;
 
     before(done => {
@@ -61,6 +61,7 @@ describe('feathers-swagger', () => {
         url: 'http://localhost:6776/docs',
         json: true
       }).then(docs => {
+        expect(docs.swagger).to.equal('2.0');
         expect(docs.info.title).to.equal('A test');
         expect(docs.info.description).to.equal('A description');
         expect(docs.paths['/messages']).to.exist;
@@ -74,6 +75,94 @@ describe('feathers-swagger', () => {
       }).then(docs => {
         const messagesIdParam = docs.paths['/messages/{id}'].get.parameters[0];
         expect(messagesIdParam.type).to.equal('string');
+      });
+    });
+
+    it('check swagger document validity', () => {
+      let swaggerSpec;
+      return rp({
+        url: 'http://localhost:6776/docs',
+        json: true
+      }).then(docs => {
+        swaggerSpec = docs;
+        return SwaggerParser.validate(docs);
+      })
+        .then(api => {
+          expect(api).to.exist;
+        })
+        .catch(error => {
+          assert.fail(`${error.message}\n\nJSON:\n${JSON.stringify(swaggerSpec, undefined, 2)}`);
+        });
+    });
+  });
+
+  describe('basic functionality with openapi v3', () => {
+    let server;
+
+    before(done => {
+      const messageService = memory();
+      messageService.docs = {
+        description: 'A service to send and receive messages',
+        definition: {
+          'type': 'object',
+          'required': [
+            'text'
+          ],
+          'properties': {
+            'text': {
+              'type': 'string',
+              'description': 'The message text'
+            },
+            'userId': {
+              'type': 'string',
+              'description': 'The id of the user that sent the message'
+            }
+          }
+        }
+      };
+
+      const app = express(feathers())
+        .configure(express.rest())
+        .configure(
+          swagger({
+            openApiVersion: 3,
+            docsPath: '/docs',
+            idType: 'string',
+            specs: {
+              info: {
+                title: 'A test',
+                description: 'A description',
+                version: '1.0.0'
+              }
+            }
+          })
+        )
+        .use('/messages', messageService);
+
+      server = app.listen(6776, () => done());
+    });
+
+    after(done => server.close(done));
+
+    it('supports basic functionality with a simple app', () => {
+      return rp({
+        url: 'http://localhost:6776/docs',
+        json: true
+      }).then(docs => {
+        expect(docs.openapi).to.equal('3.0.2');
+        expect(docs.info.title).to.equal('A test');
+        expect(docs.info.description).to.equal('A description');
+        expect(docs.paths['/messages']).to.exist;
+      });
+    });
+
+    it('supports id types in config', () => {
+      return rp({
+        url: 'http://localhost:6776/docs',
+        json: true
+      }).then(docs => {
+        const messagesIdParam = docs.paths['/messages/{id}'].get.parameters[0];
+        expect(messagesIdParam.schema.type).to.equal('string');
       });
     });